Disable Buildkite for dmd-cxx

+++ buildkite.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# Don't run Buildkite for the dmd-cxx branch
+if [ "${BUILDKITE_PULL_REQUEST_BASE_BRANCH:-master}" == "dmd-cxx" ] ; then
+    echo ""
+    return 0
+fi
+
 read -r -d '' LOAD_CI_FOLDER <<- EOM
         # just to be sure there isn't anything old left
         git clean -ffdxq .
